+++ routes/question-province/js/question-province.js
@@ -6,4 +6,14 @@ accessibleAutocomplete.enhanceSelectElement({
   selectElement: document.querySelector('#province-select'),
   showAllValues: true,
   showNoOptionsFound: false,
+  dropdownArrow: () => {
+    var altText = '';
+    if (document.documentElement.lang === 'en'){
+      altText = 'List of provinces/territories'
+    }
+    else {
+      altText = 'Liste des provinces et territoires'
+    }
+    return `<svg version="1.1" xmlns="http://www.w3.org/2000/svg" class="autocomplete__dropdown-arrow-down" alt="${altText}" focusable="false"><g stroke="none" fill="none" fill-rule="evenodd"><polygon fill="#000000" points="0 0 22 0 11 17"></polygon></g></svg>`
+  },
 })